% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_data}
\alias{get_data}
\title{Get Newly-Submitted Data For Scoring}
\usage{
get_data(cols_to_keep, metadata_filepath = NULL, test = FALSE)
}
\arguments{
\item{cols_to_keep}{Columns to keep in addition to the actions (or
multi-choice options), passed using \link[dplyr]{dplyr_tidy_select}.}

\item{metadata_filepath}{Filepath to the CSV file where the metadata is
stored.}

\item{test}{Read from the test files? Default is \code{FALSE}.}
}
\value{
A tibble.
}
\description{
Get Newly-Submitted Data For Scoring
}
\examples{
# In this example, the Sustainable Place metadata that comes shipped with the
# package would be used.
\dontrun{
get_data(cols_to_keep = timestamp:your_facility)
}
}
