% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_metadata}
\alias{get_metadata}
\title{Get Information About Each Question In The Form}
\usage{
get_metadata(metadata_filepath = NULL)
}
\arguments{
\item{metadata_filepath}{Filepath to the CSV file where the metadata is
stored.}
}
\value{
A tibble.
}
\description{
Get Information About Each Question In The Form
}
\examples{
# If no value is passed to metadata_filepath, the function returns the default
# metadata required to render the Sustainable Places Framework reports.
get_metadata()
}
