% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_new_submissions}
\alias{get_new_submissions}
\title{Create A Data Table Of Newly-Submitted Form Entries}
\usage{
get_new_submissions(test = FALSE, url = NULL)
}
\arguments{
\item{test}{Read from the test files? Default is \code{FALSE}.}

\item{url}{The URL to the Google Sheet where the data is stored.}
}
\value{
A tibble.
}
\description{
To determine whether or not new reports should be rendered, we
need to compare the data from Google Forms (live data) with the data as it
was the last time reports were rendered (snapshot data). This function
returns all the rows in the live data that are not found in the snapshot
data.
}
