% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_overall_score}
\alias{get_overall_score}
\title{Calculate Category Score For Each Pou}
\usage{
get_overall_score(data, metadata_filepath = NULL, ...)
}
\arguments{
\item{data}{Dataframe containing the score for each status ("Activator",
"Leader", or "Champion") for each Pou.}

\item{metadata_filepath}{The filepath to a CSV file containing the metadata
for the online form.}

\item{...}{Key-value pairs passed to summarise(...).}
}
\value{
A tibble.
}
\description{
Count the total number of "Leader" and "Champion" actions
selected for each Pou. If this value is at least 80\% of the total number of
"Leader" and "Champion" actions available for that Pou, assign a final
category of "Leader", otherwise assign a final category of "Activator".
}
