% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{get_question_scores}
\alias{get_question_scores}
\title{Calculate Scores Across All Questions}
\usage{
get_question_scores(data, metadata_filepath = NULL, ...)
}
\arguments{
\item{data}{A dataframe which contains, at minimum, a column for each
multi-choice selection expressed as either 1 or 0.}

\item{metadata_filepath}{The filepath to a CSV file containing the metadata
for the online form.}

\item{...}{Key-value pairs passed to select(...).}
}
\value{
A tibble.
}
\description{
For each question, calculate how many options were selected by
the user in total.
}
