% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get Live Data and Snapshot Data Together}
\usage{
get_raw_data(url = NULL)
}
\arguments{
\item{url}{The URL to the Google Sheet where the data is stored.}
}
\value{
A list of two tibbles.
}
\description{
To determine whether or not new reports should be rendered, we
need to compare the data from Google Forms (live data) with the data as it
was the last time reports were rendered (snapshot data). This function
returns a list of two tibbles, one for the live data and one for the
snapshot data.
}
