% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_scores}
\alias{get_scores}
\title{Calculate All Scores Needed For Scorecard}
\usage{
get_scores(data, metadata_path = NULL, ...)
}
\arguments{
\item{data}{A dataframe containing, ar minimum, binary values indicating
which actions were selected.}

\item{metadata_path}{Filepath to the CSV file where the metadata is stored.
By default, this is set to NULL and will return the Sustainable Places
Framework scorecard metadata.}

\item{...}{Key-value pairs determining which columns, aside from \code{id}, are
retained from \code{data}.}
}
\value{
A tibble.
}
\description{
Calculate All Scores Needed For Scorecard
}
