% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_status_scores}
\alias{get_status_scores}
\title{Calculate Scores For Each Status}
\usage{
get_status_scores(data, metadata_filepath = NULL, ...)
}
\arguments{
\item{data}{A dataframe which contains, at minimum, a column for each
multi-choice selection expressed as either 1 or 0.}

\item{metadata_filepath}{The filepath to a CSV file containing the metadata
for the online form.}

\item{...}{Key-value pairs passed to select(...).}
}
\value{
A tibble.
}
\description{
For each question, count the number of actions selected against
each status of either Activator, Champion, or Leader. The scores for the
latter two statuses determine the final category for this submission
against each Pou.
}
