% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susplasco-package.R
\docType{package}
\name{susplasco-package}
\alias{susplasco}
\alias{susplasco-package}
\title{susplasco: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auckland-Council-CC-Insights-Team/susplasco}
  \item Report bugs at \url{https://github.com/Auckland-Council-CC-Insights-Team/susplasco/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{lee.durbin@aucklandcouncil.govt.nz}

}
\keyword{internal}
