% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{get_data}
\alias{get_data}
\title{Get Newly-Submitted Data For Scoring}
\usage{
get_data(cols_to_keep, url = NULL, metadata_filepath = NULL)
}
\arguments{
\item{cols_to_keep}{Columns to keep in addition to the actions (or
multi-choice options), passed using \link[dplyr]{dplyr_tidy_select}.}

\item{url}{The URL to the Google Sheet where the data is stored. By default,
this is set to NULL and will return a fake dataset of Sustainable Places
Framework form submissions.}

\item{metadata_filepath}{Filepath to the CSV file where the metadata is
stored. By default, this is set to NULL and will return the Sustainable
Places Framework scorecard metadata.}
}
\value{
A tibble.
}
\description{
Get Newly-Submitted Data For Scoring
}
\examples{
# In this example, the Sustainable Place data and metadata that comes shipped
# with the package would be used.
\dontrun{
get_data(cols_to_keep = timestamp:your_facility)
}
}
