% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susplasco-package.R
\docType{package}
\name{susplasco-package}
\alias{susplasco}
\alias{susplasco-package}
\title{susplasco: Sustainable Places Scorecard Process Automation}
\description{
Compile a PDF report based on information received via the Sustainable Place form and save it to SharePoint.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auckland-Council-CC-Insights-Team/susplasco}
  \item Report bugs at \url{https://github.com/Auckland-Council-CC-Insights-Team/susplasco/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{lee.durbin@aucklandcouncil.govt.nz}

Other contributors:
\itemize{
  \item Auckland Council [copyright holder]
}

}
\keyword{internal}
